% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{cross_product}
\alias{cross_product}
\title{Cross Product of two vectors}
\usage{
cross_product(a, b)
}
\arguments{
\item{a}{First vector}

\item{b}{Second vector}
}
\description{
Calculates the Cross Product of two vector3D objects and returns a single vector3D object.
}
