% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datasaurus_dozen}
\alias{datasaurus_dozen}
\title{datasaurus_dozen}
\format{
A data frame with 1846 rows and 3 variables:
\itemize{
\item \strong{dataset}: Dataset to which the point belongs
\item \strong{x}: X coordinate
\item \strong{y}: Y coordinate
}
}
\usage{
datasaurus_dozen
}
\description{
A dataset demonstrating the utility of visualization. These 13 datasets are equal in standard measures: mean, standard deviation, and Pearson's correlation.
}
\keyword{datasets}
