% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datasaurus_dozen_wide}
\alias{datasaurus_dozen_wide}
\title{datasaurus_dozen_wide}
\format{
A data frame with 142 rows and 26 variables:
\itemize{
\item \strong{away_x}: x-values for the \code{away} dataset
\item \strong{away_y}: y-values for the \code{away} dataset
\item \strong{bullseye_x}: x-values for the \code{bullseye} dataset
\item \strong{bullseye_y}: y-values for the \code{bullseye} dataset
\item \strong{circle_x}: x-values for the \code{circle} dataset
\item \strong{circle_y}: y-values for the \code{circle} dataset
\item \strong{dino_x}: x-values for \code{dinosaur} dataset!
\item \strong{dino_y}: y-values for \code{dinosaur} dataset!
\item \strong{dots_x}: x-values for the \code{dots} dataset
\item \strong{dots_y}: y-values for the \code{dots} dataset
\item \strong{h_lines_x}: x-values for the \code{h_lines} dataset
\item \strong{h_lines_y}: y-values for the \code{h_lines} dataset
\item \strong{high_lines_x}: x-values for the \code{high_lines} dataset
\item \strong{high_lines_y}: y-values for the \code{high_lines} dataset
\item \strong{slant_down_x}: x-values for the \code{slant_down} dataset
\item \strong{slant_down_y}: y-values for the \code{slant_down} dataset
\item \strong{slant_up_x}: x-values for the \code{slant_up} dataset
\item \strong{slant_up_y}: y-values for the \code{slant_up} dataset
\item \strong{star_x}: x-values for the \code{star} dataset
\item \strong{star_y}: y-values for the \code{star} dataset
\item \strong{v_lines_x}: x-values for the \code{v_lines} dataset
\item \strong{v_lines_y}: y-values for the \code{v_lines} dataset
\item \strong{wide_lines_x}: x-values for the \code{wide_lines} dataset
\item \strong{wide_lines_y}: y-values for the \code{wide_lines} dataset
\item \strong{x_shape_x}: x-values for the \code{x_shape} dataset
\item \strong{x_shape_y}: y-values for the \code{x_shape} dataset
}
\usage{
datasaurus_dozen_wide
}
\description{
A dataset demonstrating the utility of visualization. These 13 datasets are equal in standard measures: mean, standard deviation, and Pearson's correlation.
}
\keyword{datasets}
