% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{del.tri}
\alias{del.tri}
\title{Implementation of the DELTRI subroutine written in R}
\usage{
del.tri(x, y = NULL, maxrange = TRUE)
}
\arguments{
\item{x}{stores the x coordinates of the points. If y is NULL, it is treated as a data.frame}

\item{y}{stores the Y coordinates of the points}

\item{maxrange}{if TRUE, normalisation is done with max(x range, y range) for both x and y values, otherwise uses x/y range for x/y normalisation}
}
\description{
Calculates the Delaunay Triangulation of the given set of points.
}
