% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{delaun}
\alias{delaun}
\title{Implementation of the DELAUN subroutine}
\usage{
delaun(norm_x, norm_y)
}
\arguments{
\item{norm_x}{Normalised X coordinates of points}

\item{norm_y}{NormalisedY coordinates of points}
}
\description{
Returns a Delaunay Triangulation but with normalised points
}
