% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DelTri}
\alias{DelTri}
\title{Implementation of the DELTRI subroutine, written in C++}
\usage{
DelTri(x_, y_ = NULL)
}
\arguments{
\item{x_}{vector of X coordinates, if y_ is NULL, x_ is a data frame}

\item{y_}{vector of Y coordinates}
}
\description{
returns an object of triad class, which can be plotted
}
