% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{edg}
\alias{edg}
\title{Implementation of the EDG subroutine written in R}
\usage{
edg(I, J, e)
}
\arguments{
\item{I}{index of triangle I}

\item{J}{index of triangle J}

\item{e}{List of adjacent triangles}
}
\description{
returns which edge of I is adjacent to J
}
