% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Edge}
\alias{Edge}
\title{Implementation of the EDG subroutine, written in C++}
\usage{
Edge(I, J, e)
}
\arguments{
\item{I}{index of triangles I}

\item{J}{index of triangle J}

\item{e}{List of adjacent triangles}
}
\description{
Returns which edge of I is adjacent to J
}
