% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetMatrix}
\alias{GetMatrix}
\title{Function to convert C++ matrices to R matrices, written in R}
\usage{
GetMatrix(m)
}
\arguments{
\item{m}{A 2D vector int object from C++ STL}
}
\description{
Converts a 2D vector int to a NumericMatrix
}
