% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hundred}
\alias{hundred}
\title{One hundred 2D points}
\format{
A data frame with 100 rows and 2 variables:
\itemize{
\item \strong{x}: x-values for the dataset
\item \strong{y}: y-values for the dataset
}
}
\usage{
hundred
}
\description{
100 randomly generated points thought runif(100)
}
\keyword{datasets}
