% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{is_left_of}
\alias{is_left_of}
\title{Checks if a point lies strictly to the left of a line segment}
\usage{
is_left_of(a, b, c)
}
\arguments{
\item{a}{initial point of a->b vector}

\item{b}{final point of a->b vector}

\item{c}{Point whose relative position is to be calculated}
}
\description{
returns TRUE if point c lies strictly to the left of a->b vector.
}
