% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{left.right}
\alias{left.right}
\title{Cross Product of AB and AP}
\usage{
left.right(Px, Py, Ax, Ay, Bx, By)
}
\arguments{
\item{Px}{X component of point P}

\item{Py}{Y component of point P}

\item{Ax}{X component of point A}

\item{Ay}{Y component of point A}

\item{Bx}{X component of point B}

\item{By}{Y component of point B}
}
\description{
Calculates cross product of AB and AP to check if P is left or right of AB
}
