% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{new_triad}
\alias{new_triad}
\title{Constructor for the triad class, written in R}
\usage{
new_triad(x, y, v)
}
\arguments{
\item{x}{X coordinates of points}

\item{y}{y coordinates of points}

\item{v}{A 2D matrix storing information of vertices that make up a triangle}
}
\description{
An object which stores the information of all vertices and triangles for the Delaunay Triangulation
}
