% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{new_vector3D}
\alias{new_vector3D}
\title{Constructor for the vector3D class}
\usage{
new_vector3D(x1 = 0, y1 = 0, z1 = 0)
}
\arguments{
\item{x1}{The x value}

\item{y1}{The y value}

\item{z1}{The z value}
}
\description{
An object of vector3D class has three values: x,y,z.
}
