% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{NewTriad}
\alias{NewTriad}
\title{Constructor for the triad class, written in C++}
\usage{
NewTriad(x, y, v)
}
\arguments{
\item{x}{Object of NumericVector class, X coordinates}

\item{y}{Object of NumericVector class, y coordinates}

\item{v}{Object of NumericMatrix class, List of triangles}
}
\description{
An object which stores the information of all vertices and triangles for the Delaunay Triangulation
}
