% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.triad}
\alias{plot.triad}
\title{plot function for objects of triad class}
\usage{
\method{plot}{triad}(x, numtext = FALSE, ...)
}
\arguments{
\item{x}{An object of triad class}

\item{numtext}{Whether points are numbered in the plot}

\item{...}{Additional arguments to pass to plot()}
}
\description{
Plots the Delaunay triangulation on passing a triad object
}
