% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{swap}
\alias{swap}
\title{Implementation of the SWAP subroutine}
\usage{
swap(x, y, P, V2, V1, V3)
}
\arguments{
\item{x}{X coordinates of points}

\item{y}{Y coordinates of points}

\item{P}{Index number of point P}

\item{V2}{Index number of point V2}

\item{V1}{Index number of point V1}

\item{V3}{Index number of point V3}
}
\description{
Checks to see if triangle P-V2-V1 and V3-V1-V2 need to swap common edge
}
