% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{swapR}
\alias{swapR}
\title{Implementation of the SWAP subroutine written in R}
\usage{
swapR(x1, y1, x2, y2, x3, y3, xp, yp)
}
\arguments{
\item{x1}{X coordinate of V1}

\item{y1}{Y coordinate of V1}

\item{x2}{X coordinate of V2}

\item{y2}{Y coordinate of V2}

\item{x3}{X coordinate of V3}

\item{y3}{Y coordinate of V3}

\item{xp}{X coordinate of P}

\item{yp}{Y coordinate of P}
}
\description{
Checks to see if triangle P-V2-V1 and V3-V1-V2 need to swap common edge
}
