% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Delaun}
\alias{Delaun}
\title{Implementation of the DELAUN subroutine}
\usage{
Delaun(norm_x, norm_y)
}
\arguments{
\item{norm_x}{Normalised X coordinates of points}

\item{norm_y}{Normalised Y coordinates of points}
}
\description{
Returns a Delaunay Triangulation but with normalised points
}
