% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{edg}
\alias{edg}
\title{Implementation of the EDG subroutine}
\usage{
edg(e, I, J)
}
\arguments{
\item{e}{List of adjacent triangles}

\item{I}{index of triangle I}

\item{J}{index of triangle J}
}
\description{
returns which edge of I is adjacent to J
}
