% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{triloc}
\alias{triloc}
\title{Check which triangles a point lies within}
\usage{
triloc(i, x, y, v, e)
}
\arguments{
\item{i}{index of the point to be located}

\item{x}{X coordinates of points}

\item{y}{Y coordinates of points}

\item{v}{vertices of triangles, columns are triangle number and rows are vertices 1, 2, and 3}

\item{e}{adjacency of triangles, columns are triangle number and rows are ids of adjacent triangles}
}
\description{
Returns the index of the triangle which contains the given point
}
